requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
click-default-group==1.2.4
colorama==0.4.6
coverage==7.4.0
Django==4.2.9
django-extensions==3.2.3
djangorestframework==3.14.0
flake8==7.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlite-fts4==1.0.3
sqlite-utils==3.36
sqlparse==0.4.4
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
